+++ logger/logger.go
@@ -1,6 +1,7 @@
 package logger
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -21,13 +22,7 @@ var (
 
 // sink interface that has to be implemented by a logging sink
 type sink interface {
-	Print(telegraf.LogLevel, time.Time, string, ...interface{})
-}
-
-// Attr represents an attribute appended to structured logging
-type Attr struct {
-	Key   string
-	Value interface{}
+	Print(telegraf.LogLevel, time.Time, string, map[string]interface{}, ...interface{})
 }
 
 // logger is the actual implementation of the telegraf logger interface
@@ -36,46 +31,25 @@ type logger struct {
 	category string
 	name     string
 	alias    string
-	suffix   string
 
-	prefix string
-
-	onError []func()
+	prefix     string
+	onError    []func()
+	attributes map[string]interface{}
 }
 
 // New creates a new logging instance to be used in models
 func New(category, name, alias string) *logger {
 	l := &logger{
-		category: category,
-		name:     name,
-		alias:    alias,
+		category:   category,
+		name:       name,
+		alias:      alias,
+		attributes: map[string]interface{}{"category": category, "plugin": name},
 	}
-	l.formatPrefix()
-
-	return l
-}
-
-// SubLogger creates a new logger with the given name added as suffix
-func (l *logger) SubLogger(name string) telegraf.Logger {
-	suffix := l.suffix
-	if suffix != "" && name != "" {
-		suffix += "."
+	if alias != "" {
+		l.attributes["alias"] = alias
 	}
-	suffix += name
 
-	nl := &logger{
-		level:    l.level,
-		category: l.category,
-		name:     l.name,
-		alias:    l.alias,
-		suffix:   suffix,
-	}
-	nl.formatPrefix()
-
-	return nl
-}
-
-func (l *logger) formatPrefix() {
+	// Format the prefix
 	l.prefix = l.category
 
 	if l.prefix != "" && l.name != "" {
@@ -88,13 +62,11 @@ func (l *logger) formatPrefix() {
 	}
 	l.prefix += l.alias
 
-	if l.suffix != "" {
-		l.prefix += "(" + l.suffix + ")"
-	}
-
 	if l.prefix != "" {
 		l.prefix = "[" + l.prefix + "] "
 	}
+
+	return l
 }
 
 // Level returns the current log-level of the logger
@@ -105,27 +77,14 @@ func (l *logger) Level() telegraf.LogLevel {
 	return instance.level
 }
 
-// SetLevel overrides the current log-level of the logger
-func (l *logger) SetLevel(level telegraf.LogLevel) {
-	l.level = &level
-}
-
-// SetLevel changes the log-level to the given one
-func (l *logger) SetLogLevel(name string) error {
-	if name == "" {
-		return nil
+// AddAttribute allows to add a key-value attribute to the logging output
+func (l *logger) AddAttribute(key string, value interface{}) {
+	// Do not allow to overwrite general keys
+	switch key {
+	case "category", "plugin", "alias":
+	default:
+		l.attributes[key] = value
 	}
-	level := telegraf.LogLevelFromString(name)
-	if level == telegraf.None {
-		return fmt.Errorf("invalid log-level %q", name)
-	}
-	l.SetLevel(level)
-	return nil
-}
-
-// Register a callback triggered when errors are about to be written to the log
-func (l *logger) RegisterErrorCallback(f func()) {
-	l.onError = append(l.onError, f)
 }
 
 // Error logging including callbacks
@@ -179,7 +138,7 @@ func (l *logger) Trace(args ...interface{}) {
 func (l *logger) Print(level telegraf.LogLevel, ts time.Time, args ...interface{}) {
 	// Check if we are in early logging state and store the message in this case
 	if instance.impl == nil {
-		instance.add(level, ts, l.prefix, args...)
+		instance.add(level, ts, l.prefix, l.attributes, args...)
 	}
 
 	// Skip all messages with insufficient log-levels
@@ -187,11 +146,34 @@ func (l *logger) Print(level telegraf.LogLevel, ts time.Time, args ...interface{
 		return
 	}
 	if instance.impl != nil {
-		instance.impl.Print(level, ts.In(instance.timezone), l.prefix, args...)
+		instance.impl.Print(level, ts.In(instance.timezone), l.prefix, l.attributes, args...)
 	} else {
 		msg := append([]interface{}{ts.In(instance.timezone).Format(time.RFC3339), " ", level.Indicator(), " ", l.prefix}, args...)
 		instance.earlysink.Print(msg...)
 	}
+}
+
+// SetLevel overrides the current log-level of the logger
+func (l *logger) SetLevel(level telegraf.LogLevel) {
+	l.level = &level
+}
+
+// SetLevel changes the log-level to the given one
+func (l *logger) SetLogLevel(name string) error {
+	if name == "" {
+		return nil
+	}
+	level := telegraf.LogLevelFromString(name)
+	if level == telegraf.None {
+		return fmt.Errorf("invalid log-level %q", name)
+	}
+	l.SetLevel(level)
+	return nil
+}
+
+// Register a callback triggered when errors are about to be written to the log
+func (l *logger) RegisterErrorCallback(f func()) {
+	l.onError = append(l.onError, f)
 }
 
 type Config struct {
@@ -199,12 +181,10 @@ type Config struct {
 	Debug bool
 	// will set the log level to ERROR
 	Quiet bool
-	//stderr, stdout, file or eventlog (Windows only)
+	// format and target of log messages
 	LogTarget string
-	// will direct the logging output to a file. Empty string is
-	// interpreted as stderr. If there is an error opening the file the
-	// logger will fall back to stderr
-	Logfile string
+	LogFormat string
+	Logfile   string
 	// will rotate when current file at the specified time interval
 	RotationInterval time.Duration
 	// will rotate when current file size exceeds this parameter.
@@ -222,6 +202,31 @@ type Config struct {
 
 // SetupLogging configures the logging output.
 func SetupLogging(cfg *Config) error {
+	// Issue deprecation warning for option
+	switch cfg.LogTarget {
+	case "":
+		// Best-case no target set or file already migrated...
+	case "stderr":
+		msg := "Agent setting %q is deprecated, please leave %q empty and remove this setting!"
+		deprecation := "The setting will be removed in v1.40.0."
+		log.Printf("W! "+msg+" "+deprecation, "logtarget", "logfile")
+		cfg.Logfile = ""
+	case "file":
+		msg := "Agent setting %q is deprecated, please just set %q and remove this setting!"
+		deprecation := "The setting will be removed in v1.40.0."
+		log.Printf("W! "+msg+" "+deprecation, "logtarget", "logfile")
+	case "eventlog":
+		msg := "Agent setting %q is deprecated, please set %q to %q and remove this setting!"
+		deprecation := "The setting will be removed in v1.40.0."
+		log.Printf("W! "+msg+" "+deprecation, "logtarget", "logformat", "eventlog")
+		if cfg.LogFormat != "" && cfg.LogFormat != "eventlog" {
+			return errors.New("contradicting setting between 'logtarget' and 'logformat'")
+		}
+		cfg.LogFormat = "eventlog"
+	default:
+		return fmt.Errorf("invalid deprecated 'logtarget' setting %q", cfg.LogTarget)
+	}
+
 	if cfg.Debug {
 		cfg.logLevel = telegraf.Debug
 	}
@@ -236,8 +241,8 @@ func SetupLogging(cfg *Config) error {
 		cfg.InstanceName = "telegraf"
 	}
 
-	if cfg.LogTarget == "" || cfg.LogTarget == "file" && cfg.Logfile == "" {
-		cfg.LogTarget = "stderr"
+	if cfg.LogFormat == "" {
+		cfg.LogFormat = "text"
 	}
 
 	// Get configured timezone
@@ -250,13 +255,12 @@ func SetupLogging(cfg *Config) error {
 		return fmt.Errorf("setting logging timezone failed: %w", err)
 	}
 
-	// Get the logging factory
-	creator, ok := registry[cfg.LogTarget]
-	if !ok {
-		return fmt.Errorf("unsupported log target: %s, using stderr", cfg.LogTarget)
+	// Get the logging factory and create the root instance
+	creator, found := registry[cfg.LogFormat]
+	if !found {
+		return fmt.Errorf("unsupported log-format: %s", cfg.LogFormat)
 	}
 
-	// Create the root logging instance
 	l, err := creator(cfg)
 	if err != nil {
 		return err
@@ -268,7 +272,8 @@ func SetupLogging(cfg *Config) error {
 	}
 
 	// Update the logging instance
-	instance.switchSink(l, cfg.logLevel, tz, cfg.LogTarget == "stderr")
+	skipEarlyLogs := cfg.LogFormat == "text" && cfg.Logfile == ""
+	instance.switchSink(l, cfg.logLevel, tz, skipEarlyLogs)
 
 	return nil
 }