+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 
 ### Important Changes
 
+- This release contains a logging overhaul as well as some new features for
+  logging (see PRs [#15556](https://github.com/influxdata/telegraf/pull/15556),
+  [#15629](https://github.com/influxdata/telegraf/pull/15629),
+  [#15677](https://github.com/influxdata/telegraf/pull/15677),
+  [#15695](https://github.com/influxdata/telegraf/pull/15695) and
+  [#15751](https://github.com/influxdata/telegraf/pull/15751)).
+  As a consequence the redunant `logtarget` setting is deprecated, `stderr` is
+  used if no `logfile` is provided, otherwise messages are logged to the given
+  file. For using the Windows `eventlog` set `logformat = "eventlog"`!
 - This release contains a change in json_v2 parser config parsing -
   if the config is empty (not define any rules), initialization will fail
   (see PR [#15844](https://github.com/influxdata/telegraf/pull/15844)).